% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cord.R
\name{cord}
\alias{cord}
\title{Model based ordination with gaussian copulas}
\usage{
cord(obj, nlv = 2, n.samp = 500, seed = 1)
}
\arguments{
\item{obj}{object of class manyglm, or manyany with "clm"}

\item{nlv}{number of latent variables (default is 2)}

\item{n.samp}{number of residuals used for importance sampling (default is 500)}
}
\description{
Model based ordination with gaussian copulas
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider.mod=manyglm(abund~X)
spid.lv=cord(spider.mod)
plot(spid.lv)


}
