% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cord.R
\name{ordiplot.cord}
\alias{ordiplot.cord}
\title{Plot graph of species intearactions.}
\usage{
\method{ordiplot}{cord}(obj, biplot = FALSE, ...)
}
\arguments{
\item{obj}{is a cord object, e.g. from output of \code{cord}}
}
\value{
an ordination plot.
}
\description{
Plot graph of species intearactions.
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider_mod=manyglm(abund~1)
spid_lv=cord(spider_mod)
#colour sites accoring to second column of x (bare sand)
cols=ifelse(spider$x[,2]>0,"black","red")
ordiplot(spid_lv,biplot = TRUE,col=cols)
}
