% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cord.R
\name{ordiplot}
\alias{ordiplot}
\title{Plots latent variables and their corresponding coefficients (biplot).}
\usage{
ordiplot(x, ...)
}
\arguments{
\item{obj}{is a cord object, e.g. from output of \code{cord}}

\item{biplot}{\code{TRUE} if both latent variables and their coefficients are plotted, \code{FALSE} if only latent variables}
}
\value{
an ordination plot.
}
\description{
Plots latent variables and their corresponding coefficients (biplot).
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider_mod=manyglm(abund~1)
spid_lv=cord(spider_mod)
#colour sites accoring to second column of x (bare sand)
cols=ifelse(spider$x[,2]>0,"black","red")
ordiplot(spid_lv,biplot = TRUE,col=cols)
}
