% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saint.R
\name{plot.saint}
\alias{plot.saint}
\title{Plot graph of species intearactions.}
\usage{
\method{plot}{saint}(obj, P = NULL, ...)
}
\arguments{
\item{obj}{is a saint object, e.g. from output of saint.}

\item{P}{locations of graph nodes, if NULL (default) these are generated with a Fruchterman Reingold algorithm.}

\item{...}{other parameters to be passed through to plotting functions, in particular \code{pad}, the amount to pad the plotting range is useful if labels are being clipped.}
}
\value{
a plot of species associations after accounting for the effect of all other species, positive/negative are blue/pink.
The matrix of node positions (\code{P}) is returned silently.
}
\description{
Plot graph of species intearactions.
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider_mod=manyglm(abund~1)
spid_graph=saint(spider_mod)
plot(spid_graph)
}
