% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedsdm_s3.R
\name{plot.stackedsdm}
\alias{plot.stackedsdm}
\title{Plot residuals of stackedsdm.}
\usage{
\method{plot}{stackedsdm}(x, ...)
}
\arguments{
\item{x}{is a stackedsdm object.}

\item{...}{not used}
}
\description{
Plot residuals of stackedsdm.
}
\examples{
X <- as.data.frame(spider$x)
abund <- spider$abund
spider_mod <- stackedsdm(abund,~1, data = X) 
plot(spider_mod)
}
