% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cord_s3.R
\name{print.cord}
\alias{print.cord}
\title{Print function for cord object}
\usage{
\method{print}{cord}(x, ...)
}
\arguments{
\item{x}{is a cord object, e.g. from output of \code{\link{cord}}.}

\item{...}{not used}
}
\description{
Print function for cord object
}
\examples{
X <- as.data.frame(spider$x)
abund <- spider$abund
spider_mod <- stackedsdm(abund,~1, data = X) 
spid_lv=cord(spider_mod)
print(spid_lv)
}
\seealso{
\code{\link{cord}}
}
