% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saint.R
\name{saint}
\alias{saint}
\title{Fit gaussian copula graphical lasso, chose lambda with BIC.}
\usage{
saint(manyglm.obj, n.lambda = 100, n.samp = 200)
}
\arguments{
\item{manyglm.obj}{object of class manyglm.}

\item{n.lambda}{number of values of shrinkage parameter lambda for model selection.}

\item{n.samp}{number of residuals used for importance sampling.}
}
\value{
best.graph a list with the paramaters for the best graphical model

all.graphs a list with likelihood and BIC for all models along lambda path
}
\description{
Fit gaussian copula graphical lasso, chose lambda with BIC.
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider.mod=manyglm(abund~X)
spid.graph=saint(spider.mod)
plot(spid.graph)


}
