% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cord.R
\name{simulate.cord}
\alias{simulate.cord}
\title{Simulates new data from a given cord object}
\usage{
\method{simulate}{cord}(fit, nsim = 1, seed = NULL)
}
\arguments{
\item{nsim}{number of simulations, defaults to 1}

\item{seed}{random number seed}

\item{obj}{is a cord object, e.g. from output of \code{cord}}
}
\description{
Simulates new data from a given cord object
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider_mod=manyglm(abund~1)
spid_lv=cord(spider_mod)
simulate(spid_lv)
}
