% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgr_s3.R
\name{summary.cgr}
\alias{summary.cgr}
\title{Summary function for cgr object}
\usage{
\method{summary}{cgr}(object, ...)
}
\arguments{
\item{object}{is a cgr object, e.g. from output of \code{\link{cgr}}.}

\item{...}{not used}
}
\description{
Summary function for cgr object
}
\examples{
\donttest{
abund <- spider$abund[,1:5]
spider_mod <- stackedsdm(abund,~1, data = spider$x, ncores=2) 
spid_graph=cgr(spider_mod)
summary(spid_graph)
}
}
\seealso{
\code{\link{cgr}}
}
