% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cord_s3.R
\name{summary.cord}
\alias{summary.cord}
\title{Summary function for cgr object}
\usage{
\method{summary}{cord}(object, ...)
}
\arguments{
\item{object}{is a cord object, e.g. from output of \code{\link{cgr}}.}

\item{...}{not used}
}
\description{
Summary function for cgr object
}
\examples{
abund <- spider$abund[,1:5]
spider_mod <- stackedsdm(abund,~1, data = spider$x, ncores=2) 
spid_lv=cord(spider_mod)
summary(spid_lv)
}
\seealso{
\code{\link{cord}}
}
