% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cord.R
\name{plot.cord}
\alias{plot.cord}
\title{Plots an ordination plot of latent variables and their corresponding coefficients (biplot).}
\usage{
\method{plot}{cord}(
  obj,
  biplot = FALSE,
  site.col = "black",
  sp.col = "blue",
  alpha = 0.9,
  arrow = TRUE,
  ...
)
}
\arguments{
\item{obj}{is a cord object, e.g. from output of \code{cord}}

\item{biplot}{\code{TRUE} if both latent variables and their coefficients are plotted, \code{FALSE} if only latent variables}

\item{site.col}{site number colour (default is black), vector of length equal to the number of sites}

\item{sp.col}{species name colour (default is blue), vector of length equal to the number of sites (if arrow=TRUE)}

\item{alpha}{scaling factor for ratio of scores to loadings (default is 0.9)}

\item{arrow}{should arrows be plotted for species loadings (default is TRUE)}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
an ordination plot.
}
\description{
Plots an ordination plot of latent variables and their corresponding coefficients (biplot).
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider_mod=manyglm(abund~1)
spid_lv=cord(spider_mod)
#colour sites accoring to second column of x (bare sand)
cols=ifelse(spider$x[,2]>0,"black","red")
plot(spid_lv,biplot = TRUE,site.col=cols)
}
