% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saint_plot.R
\name{plot.saint}
\alias{plot.saint}
\title{Plot graph of species intearactions.}
\usage{
\method{plot}{saint}(obj, P = NULL)
}
\arguments{
\item{obj}{is a mvabund object, e.g. from output of saint}
}
\value{
a plot of species interactions, positive/negative interactions are blue/pink,
}
\description{
Plot graph of species intearactions.
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider.mod=manyglm(abund~X)
spid.graph=saint(spider.mod)
plot(spid.graph)


}
