% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saint.R
\name{saint}
\alias{saint}
\title{Fit gaussian copula graphical lasso, chose lambda with BIC.}
\usage{
saint(obj, lambda = NULL, n.lambda = 100, n.samp = 500, seed = 1)
}
\arguments{
\item{obj}{object of class manyglm, or manyany with "clm"}

\item{lambda}{values of shrinkage parameter lambda for model selection (optional)}

\item{n.lambda}{number of values of shrinkage parameter lambda for model selection, ignored if lambda suplied}

\item{n.samp}{number of residuals used for importance sampling (optional)}
}
\value{
best.graph a list with the paramaters for the 'best' graphical model, chosen by BIC

all.graphs a list with likelihood and BIC for all models along lambda path
}
\description{
Fit gaussian copula graphical lasso, chose lambda with BIC.
}
\examples{
data(spider)
abund <- mvabund(spider$abund)
X <- spider$x
spider.mod=manyglm(abund~X)
spid.graph=saint(spider.mod)
plot(spid.graph)


}
